requirements.txt
influxdb_client==1.30.0
python-dotenv==1.0.0
requests==2.27.1
